完善timing部分 1）补充about界面缺失标签； 2）修复timing界面slack显示问题，从整数修改为小数； 3）自动加载runs/impl/下的timing json文件；

+++ src/component/timing/DesignTimingSummaryWidget.cpp
@@ -1,9 +1,9 @@
 #include "DesignTimingSummaryWidget.h"
 
-DesignTimingSummaryWidget::DesignTimingSummaryWidget(const int &worstNegativeSlack,
-                                                     const int &totalNegativeSlack,
-                                                     const int &worstHoldSlack,
-                                                     const int &totalHoldSlack,
+DesignTimingSummaryWidget::DesignTimingSummaryWidget(const float &worstNegativeSlack,
+                                                     const float &totalNegativeSlack,
+                                                     const float &worstHoldSlack,
+                                                     const float &totalHoldSlack,
                                                      QWidget *parent)
     : QWidget(parent)
 {
